+++ src/main/java/dev/willyelton/crystal_tools/common/events/PlayerEvents.java
@@ -1,9 +1,14 @@
 package dev.willyelton.crystal_tools.common.events;
 
 import dev.willyelton.crystal_tools.CrystalTools;
+import dev.willyelton.crystal_tools.Registration;
+import dev.willyelton.crystal_tools.common.capability.Capabilities;
+import dev.willyelton.crystal_tools.common.capability.Levelable;
 import dev.willyelton.crystal_tools.common.components.DataComponents;
 import dev.willyelton.crystal_tools.common.inventory.CrystalBackpackInventory;
 import dev.willyelton.crystal_tools.common.levelable.CrystalBackpack;
+import dev.willyelton.crystal_tools.utils.InventoryUtils;
+import dev.willyelton.crystal_tools.utils.ToolUtils;
 import net.minecraft.util.TriState;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.bus.api.SubscribeEvent;
@@ -17,6 +22,16 @@ public class PlayerEvents {
     @SubscribeEvent
     public static void handleItemPickupEvent(ItemEntityPickupEvent.Pre event) {
         if (event.getItemEntity().hasPickUpDelay()) return;
+        if (event.getItemEntity().getData(Registration.MAGNET_ITEM.get()) && !event.getPlayer().equals(event.getItemEntity().getOwner())) {
+            List<ItemStack> magnetStacks = InventoryUtils.findAll(event.getPlayer(), stack -> stack.is(Registration.CRYSTAL_MAGNET.get()));
+            magnetStacks.forEach(stack -> {
+                Levelable levelable = stack.getCapability(Capabilities.ITEM_SKILL, null);
+                if (levelable != null) {
+                    levelable.addExp(event.getPlayer().level(), event.getPlayer().getOnPos(), event.getPlayer(), 1);
+                }
+            });
+        }
+
         List<ItemStack> backpackStacks = CrystalBackpack.findBackpackStacks(event.getPlayer());
         ItemStack stackToInsert = event.getItemEntity().getItem();
         ItemStack original = event.getItemEntity().getItem().copy();
@@ -27,9 +42,9 @@ public class PlayerEvents {
             }
 
             CrystalBackpackInventory inventory = CrystalBackpack.getInventory(backpackStack);
-            List<ItemStack> filterStacks = CrystalBackpack.getFilterItems(backpackStack);
+            List<ItemStack> filterStacks = ToolUtils.getFilterItems(backpackStack);
 
-            if (!shouldPickup(backpackStack, stackToInsert, filterStacks)) {
+            if (!ToolUtils.matchesFilter(backpackStack, stackToInsert, filterStacks)) {
                 continue;
             }
 
@@ -51,21 +66,5 @@ public class PlayerEvents {
             event.setCanPickup(TriState.DEFAULT);
 //            event.setCanceled(true);
         }
-    }
-
-    private static boolean shouldPickup(ItemStack backpackStack, ItemStack pickupStack, List<ItemStack> filter) {
-        // TODO: matching modes (respect nbt)
-        // If there are no filters, default to pickup
-        if (backpackStack.getOrDefault(DataComponents.FILTER_CAPACITY, 0) == 0) {
-            return true;
-        }
-        boolean whiteList = backpackStack.getOrDefault(DataComponents.WHITELIST, true);
-        for (ItemStack filterStack : filter) {
-            if (filterStack.is(pickupStack.getItem())) {
-                return whiteList;
-            }
-        }
-
-        return !whiteList;
     }
 }