+++ src/main/java/dev/willyelton/crystal_tools/common/levelable/tool/CrystalRocket.java
@@ -29,7 +29,6 @@ import java.util.function.Consumer;
 public class CrystalRocket extends LevelableTool {
     public CrystalRocket(Item.Properties properties) {
         super(properties
-                .repairable(CrystalToolsTags.REPAIRS_CRYSTAL)
                 .durability(100));
     }